MixerGUI, AnimationPanel improvements
Values shown in the MixerGUI were updated to provide more useful information. Each "channel" displays its individual weight with a max of 1. To find the denominator for the weighting, use sum of all the channels.

The Master slider now displays a value that is used as a multiplier. If the Master slider is at "1", there is no effect--the channel values pass through. If the master is set at, for example, 2, all the individual channels will be doubled prior to being mapped to a color value. Noise values are still subject to "ring" or "clamp" transforms if they are greater than 1 or less than 0, which now more logically relates to the text field values.

For the Animation Panel: added crude tabbing to the input fields, and fixed bug so that now the animation stops when exiting the panel. Also tried to make the button logic more consistent. For example, changing the number of frames or reloading data will reset the view slider to 0.

Some changes were started to the tutorial pages, to be more consistent with the new mixer GUI, but I stopped when realizing the tutorial is being suppressed and is in need of a major overhaul.