Reboot revisions
Many revisions.
(1) Most significant: MixerModel now holds the "all" slider value and the TextureFunction does a multiplication prior to the main loop to determine the individual channel weights.
(2) AnimationPanel easier to use, including getting rid of the modal, and having changes to the general settings trigger a requirement for a reload.
(3) Order of panels on main panel revised to go with the flow of the texture calculations. NOTE: I put the prefilter (ring/clamp) with the ColorMaps, but the model value remains with the MixerModel.
(4) Added ability to choose NoiseEngine, and added OpenSimplexNoise as the default.
(5) Some tinkering with the gallery as the lattices for OpenSimplexNoise seem to me to have a wider spacing than Perlin's Simplex Noise.
(6) A lot of work has been done on the README.md, but I am not quite ready to commit it.