New README, elim tutorial, some renames
Biggest change here is the new README which comes with some graphics, and a link to a downloadable source that can be compiled and run.
Some significant renames, including the LaunchSivi (holding our main() method) and its parent folder mean that all classes are affected (in the imports line). The "tutorial" section with all its graphics has been deleted ("archived" elsewhere).